requirements.txt
Flask==2.3.2
pytesseract==0.3.10
Pillow==8.4.0
pytest==6.2.5
pytest-cov==3.0.0
requests==2.28.2
